+++ include/retdec/cpdetect/settings.h
@@ -24,7 +24,8 @@ const std::size_t EP_BYTES_SIZE = 50;
 const std::set<std::string> EXTERNAL_DATABASE_SUFFIXES =
 {
 	".yar",
-	".yara"
+	".yara",
+	".yarac"
 };
 
 const std::string YARA_RULES_PATH = "../share/retdec/support/generic/yara_patterns/tools/";